requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.11
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
decouple==0.0.7
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-allauth==0.48.0
django-betterforms==1.2
django-countries==7.2.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-dotenv==1.4.2
django-extensions==3.1.5
django-heroku==0.3.1
django-seed==0.3.1
django-werkzeug==1.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djlint==0.7.4
djongo==1.3.6
drf-excel==2.1.0
drf-renderer-xlsx==1.0.0
drf-writable-nested==0.6.3
environ==1.0
et-xmlfile==1.1.0
Faker==13.3.4
filelock==3.6.0
flake8==4.0.1
gunicorn==20.1.0
identify==2.4.12
idna==3.3
importlib-metadata==4.11.1
isort==5.10.1
mccabe==0.6.1
mysqlclient==2.1.0
nodeenv==1.6.0
oauthlib==3.2.0
openpyxl==3.0.9
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pre-commit==2.18.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pymongo==3.12.3
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
razorpay==1.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.2.4
stripe==2.66.0
toml==0.10.2
tomlkit==0.8.0
toposort==1.7
tqdm==4.62.3
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.14.1
Werkzeug==2.0.3
whitenoise==6.0.0
zipp==3.7.0
